[submitter] Update tractor submitter
